adds headers and styles things a bit
